HH-218702 move DataSourceContext to nab-jdbc module
