+++ nab-hibernate-persistence-provider/src/main/java/ru/hh/nab/hibernate/routing/DataSourceRoutingFilter.java
@@ -11,8 +11,8 @@ import java.io.IOException;
 import org.glassfish.jersey.uri.UriComponent;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import ru.hh.nab.hibernate.transaction.DataSourceContext;
 import ru.hh.nab.jdbc.common.DataSourceType;
+import ru.hh.nab.jdbc.routing.DataSourceContext;
 import ru.hh.nab.jdbc.routing.DataSourceContextUnsafe;
 
 public class DataSourceRoutingFilter implements Filter {
