+++ nab-jdbc/pom.xml
@@ -22,6 +22,10 @@
             <groupId>ru.hh.nab</groupId>
             <artifactId>nab-jdbc-common</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-jdbc</artifactId>
+        </dependency>
     </dependencies>
 
 </project>
